Fix PrimUnlifted instances for small arrays
When using GHC versions before 7.10, `SmallArray` is a newtype
wrapper around `Array`, and similarly `SmallMutableArray` is a
newtype wrapper around `MutableArray`. So the `PrimUnlifted`
instances need to go through two layers of constructors to get
to an unlifted type. Previously we failed to do that, leading
to invalid and dangerous coercions.