Set the stage for UnliftedArray# (#196)
The `PrimUnlifted` class, used to implement `UnliftedArray`
operations, is currently based on an `ArrayArray#` representation
of unlifted arrays. But that's set to change, hopefully soon.
I don't think it's too soon to set up the machinery necessary to
allow users to write code *today* that will work once `ArrayArray#`
is gone.

The key here is to add an associated type family to `PrimUnlifted`
that ties each `LiftedRep` type to its `UnliftedRep` counterpart,
and methods that convert back and forth.