Remove Cabal flag hack in primitive-benchmarks
It turns out that we can get away without needing an ugly Cabal flag
in `primitive-benchmarks` by simply using `haskell-ci` more
intelligently. I only added this flag in the first place since I was
under the false impression that `haskell-ci` would _always_ pass the
`--enable-benchmarks` flag if there was a benchmark suite, but I have
since learned that that is not true.